+++ packages/cli/src/compile/compile.ts
@@ -69,7 +69,11 @@ export const loadTransformOptions = async (
     const path = await resolveSpecifierPath(pattern, opts.repoDir, log);
     if (path) {
       try {
-        exports = await preloadAdaptorExports(path, log);
+        exports = await preloadAdaptorExports(
+          path,
+          opts.useAdaptorsMonorepo,
+          log
+        );
       } catch (e) {
         log.error(`Failed to load adaptor typedefs from path ${path}`);
         log.error(e);